requirements.txt
pypolycontain==1.4
osmium==3.6.0
pyproj==3.5.0
alphashape==1.3.1
descartes==1.1.0
scikit-learn==0.24.1
shapely==1.8.5
matplotlib==3.3.4
numpy==1.24.4
pandas==2.0.3
tqdm==4.66.1
black==24.2.0
pytorch 
torchvision